Each connection gets its own struct
This is done so we can later be able to hold state (like no. of bytes
read) and properly process each incoming message.

For example we now assume that recv(2) will get the whole message in a
single read, which is not always the case. By keeping a separate buffer
per connection we can properly process incoming messages concurrently.

Also extract some stuff to a header file.